requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
attributee=0.1.8=pypi_0
attrs=23.1.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
bidict=0.22.1=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.01.10=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.0.1=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py38hdb19cb5_0
cudatoolkit=11.3.1=h9edb442_10
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.32=py38h6a678d5_0
dbus=1.13.18=hb2f20db_0
dominate=2.8.0=pypi_0
einops=0.6.0=pypi_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
gdown=4.6.2=pypi_0
giflib=5.2.1=h5eee18b_1
glib=2.69.1=he621ea3_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.12.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.25.0=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=h7f98852_2003
jpeg=9e=h166bdaf_1
jpeg4py=0.1.4=pypi_0
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7f98852_1001
lazy-object-proxy=1.9.0=pypi_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=2.2.1=h9c3ff4c_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.7=h7f98852_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=h21135ba_2
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuuid=1.41.5=h5eee18b_0
libuv=1.43.0=h7f98852_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvmlite=0.40.0=pypi_0
lvis=0.5.3=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.2=py38h06a4308_0
matplotlib-base=3.6.2=py38h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
networkx=3.0=pypi_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.57.0=pypi_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.7.0.68=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1s=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=22.0=py38h06a4308_0
pandas=1.5.2=py38h417a72b_0
pcre=8.45=h295c915_0
phx-class-registry=4.0.6=pypi_0
pillow=8.3.2=py38h8e6f84c_0
pip=22.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
ply=3.11=py38_0
protobuf=3.20.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pypi_0
pylatex=1.4.1=pypi_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.10.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py38h06a4308_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.10.0=pypi_0
segment-anything=1.0=dev_0
setuptools=65.6.3=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.2.post1=pypi_0
spatial-correlation-sampler=0.4.0=pypi_0
sqlite=3.40.1=h5082296_0
tb-nightly=2.13.0a20230212=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6=pypi_0
tifffile=2023.2.3=pypi_0
tikzplotlib=0.10.1=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.10.0=py38_cu113
torchvision=0.11.0=py38_cu113
tornado=6.2=py38h5eee18b_0
tqdm=4.64.1=py38h06a4308_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.14=pypi_0
visdom=0.2.4=pypi_0
vot-toolkit=0.5.3=pypi_0
vot-trax=3.0.2=pypi_0
webcolors=1.12=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
zipp=3.13.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
